rename test directory
