+++ src/test/java/de/usd/cstchef/operations/arithmetic/MedianTest.java
@@ -10,7 +10,7 @@ import de.usd.cstchef.operations.Operation.OperationInfos;
 import de.usd.cstchef.operations.Operation;
 import de.usd.cstchef.operations.OperationCategory;
 import de.usd.cstchef.operations.arithmetic.Median;
-import de.usd.cstchef.testutils.UnitTestObjectFactory;
+import de.usd.cstchef.utils.UnitTestObjectFactory;
 
 @OperationInfos(name = "Test", category = OperationCategory.ARITHMETIC, description = "Test class")
 public class MedianTest extends Median
