+++ src/test/java/de/usd/cstchef/operations/arithmetic/MultiplyListTest.java
@@ -10,7 +10,7 @@ import de.usd.cstchef.operations.Operation.OperationInfos;
 import de.usd.cstchef.operations.Operation;
 import de.usd.cstchef.operations.OperationCategory;
 import de.usd.cstchef.operations.arithmetic.MultiplyList;
-import de.usd.cstchef.testutils.UnitTestObjectFactory;
+import de.usd.cstchef.utils.UnitTestObjectFactory;
 
 @OperationInfos(name = "Test", category = OperationCategory.ARITHMETIC, description = "Test class")
 public class MultiplyListTest extends MultiplyList
